Låna/returnera flyttat
Besökare kan låna på ID

Bibliotekarie kan registrera återlämning
