+++ src/main/java/com/newtongroup/library/Controller/LibrarianController.java
@@ -1,8 +1,7 @@
 package com.newtongroup.library.Controller;
 
-import com.newtongroup.library.Entity.User;
-import com.newtongroup.library.Repository.UserRepository;
-import com.newtongroup.library.Repository.VisitorRepository;
+import com.newtongroup.library.Entity.*;
+import com.newtongroup.library.Repository.*;
 import com.newtongroup.library.Utils.HeaderUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,6 +10,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.security.Principal;
+import java.sql.Date;
+import java.util.Calendar;
 
 @Controller
 @RequestMapping("/librarian")
@@ -21,6 +22,18 @@ public class LibrarianController {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private BookRepository bookrepository;
+
+    @Autowired
+    private BookLoanRepository bookLoanRepository;
+
+    @Autowired
+    private EbookLoanRepository ebookLoanRepository;
+
+    @Autowired
+    private EBookRepository eBookRepository;
 
     private String header = "librarian/bootstrapheader.html";
 
@@ -53,4 +66,53 @@ public class LibrarianController {
 
         return "admin/delete-confirmation";
     }
+
+    @RequestMapping("/input-book-to-return")
+    private String prepareToReturnBook(Model theModel, Principal principal) {
+        System.out.println(principal.getName());
+        theModel.addAttribute("header", HeaderUtils.getHeaderString(userRepository.findByUsername(principal.getName())));
+        return "loan/return-book";
+    }
+
+    @RequestMapping("/return-book")
+    private String returnBook(@RequestParam(name="bookId", required = false) Long bookId,
+                              @RequestParam(name="eBookId", required = false) Long eBookId,
+                              Model theModel, Principal principal) {
+
+        theModel.addAttribute("header", HeaderUtils.getHeaderString(userRepository.findByUsername(principal.getName())));
+
+
+        if(bookId != null) {
+            Book bookToReturn = bookrepository.findById(bookId).orElse(null);
+            BookLoan loan = bookLoanRepository.findByBookAndIsBookReturned(bookToReturn, false);
+
+            if(loan == null) {
+                return "error/book-or-no-active-librarycard";
+            }
+
+            bookToReturn.setAvailable(true);
+            loan.setDateReturned(new Date(Calendar.getInstance().getTime().getTime()));
+            loan.setBookReturned(true);
+            bookLoanRepository.save(loan);
+
+        } else if ( eBookId != null) {
+            EBook bookToReturn = eBookRepository.findById(eBookId).orElse(null);
+            EbookLoan loan = ebookLoanRepository.findByEbookAndIsEbookReturned(bookToReturn, false);
+
+            if(loan == null) {
+                return "error/book-or-no-active-librarycard";
+            }
+
+            bookToReturn.setAvailable(true);
+            loan.setDateReturned(new Date(Calendar.getInstance().getTime().getTime()));
+            loan.setEbookReturned(true);
+            ebookLoanRepository.save(loan);
+
+        } else {
+            return "error/book-or-no-active-librarycard";
+        }
+
+        return "loan/return-success";
+    }
+
 }
