+++ src/main/java/com/newtongroup/library/Controller/LoanController.java
@@ -35,6 +35,13 @@ public class LoanController {
     @Autowired
     private EBookRepository eBookRepository;
 
+    @RequestMapping("/")
+    public String loanhome(Model theModel, Principal principal){
+        theModel.addAttribute("header", HeaderUtils.getHeaderString(userRepository.findByUsername(principal.getName())));
+
+        return "loan/register-book";
+    }
+
     @RequestMapping("/register-loan")
     public String registerLoan(@RequestParam(value = "bookId", required = false) Long bookId,
                                @RequestParam(name="eBookId", required = false) Long eBookId,
@@ -70,53 +77,7 @@ public class LoanController {
 
     }
 
-    @RequestMapping("/input-book-to-return")
-    private String prepareToReturnBook(Model theModel, Principal principal) {
-        System.out.println(principal.getName());
-        theModel.addAttribute("header", HeaderUtils.getHeaderString(userRepository.findByUsername(principal.getName())));
-        return "loan/return-book";
-    }
 
-    @RequestMapping("/return-book")
-    private String returnBook(@RequestParam(name="bookId", required = false) Long bookId,
-                              @RequestParam(name="eBookId", required = false) Long eBookId,
-                              Model theModel, Principal principal) {
-
-        theModel.addAttribute("header", HeaderUtils.getHeaderString(userRepository.findByUsername(principal.getName())));
-
-
-        if(bookId != null) {
-            Book bookToReturn = bookrepository.findById(bookId).orElse(null);
-            BookLoan loan = bookLoanRepository.findByBookAndIsBookReturned(bookToReturn, false);
-
-            if(loan == null) {
-                return "error/book-or-no-active-librarycard";
-            }
-
-            bookToReturn.setAvailable(true);
-            loan.setDateReturned(new Date(Calendar.getInstance().getTime().getTime()));
-            loan.setBookReturned(true);
-            bookLoanRepository.save(loan);
-
-        } else if ( eBookId != null) {
-            EBook bookToReturn = eBookRepository.findById(eBookId).orElse(null);
-            EbookLoan loan = ebookLoanRepository.findByEbookAndIsEbookReturned(bookToReturn, false);
-
-            if(loan == null) {
-                return "error/book-or-no-active-librarycard";
-            }
-
-            bookToReturn.setAvailable(true);
-            loan.setDateReturned(new Date(Calendar.getInstance().getTime().getTime()));
-            loan.setEbookReturned(true);
-            ebookLoanRepository.save(loan);
-
-        } else {
-            return "error/book-or-no-active-librarycard";
-        }
-
-        return "loan/return-success";
-    }
 
     private boolean doesVisitorHaveActiveLibraryCard(Visitor visitor) {
         return visitor.getActiveLibraryCard() != null;
