+++ src/main/resources/templates/librarian/bootstrapheader.html
@@ -59,6 +59,10 @@
                 <li class="nav-item">
                     <a class="nav-link" href="/register-visitor/">Lägg till besökare</a>
                 </li>
+
+                <li class="nav-item">
+                    <a class="nav-link" href="/librarian/input-book-to-return">Återlämning</a>
+                </li>
                 <form class="form-inline" th:action="@{/search}" method="post">
                     <input class="searchLibrarian form-control mr-sm-2" name="search" type="search" placeholder="bok / seminarium" aria-label="Search">
                     <button class="btn2" type="submit">Sök</button>
