+++ src/main/resources/templates/visitor/bootstrapheader.html
@@ -35,8 +35,9 @@
                     <a class="nav-link" href="/visitor/my-loans">Mina Lånade böcker</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/loan/input-book-to-return">Lämna tillbaka böcker</a>
+                    <a class="nav-link" href="/loan/">Låna bok</a>
                 </li>
+
                 <form class="form-inline" th:action="@{/search}" method="post">
                     <input class="searchVisitor form-control mr-sm-2" name="search" type="search" placeholder="bok / seminarium" aria-label="Search">
                     <button class="btn2" type="submit">Sök</button>
